Re-implemented QCL
Implemented QCL-QPE using the existing backbone so that it works, and is
parametrizable in depth. Two implementation paths were tried out - the
first one was supposed to use the Operator class of Qiskit, and the
second one replaced it by using a function written by me.

Due to unclear dimensionality and controlling of the Operator class, with no
elaboration in the Qiskit documentation, the first implementation is not
complete. The second one is almost complete, with extension to
higher-dimension matrices still remaining.

The function mapAtoU3() is also updated, eliminating some bugs.